Big int in Cassandra casts to int64 Cast big int in Cassandra to int64 upon read
Summary:
version is store as big int in Cassandra,
rocql would cast big int to int64. Wrong type casting would cause panic.

Reviewers: apoorvaj, #peloton

Reviewed By: apoorvaj, #peloton

Subscribers: jenkins

Maniphest Tasks: T1794671

Differential Revision: https://code.uberinternal.com/D1771239

+++ storage/cassandra/store.go
@@ -443,7 +443,9 @@ func (s *Store) getMaxJobVersion(ctx context.Context, id *peloton.JobID) (uint64
 	log.Debugf("max version: %v", allResults)
 	for _, value := range allResults {
 		for _, max := range value {
-			return max.(uint64), nil
+			// version is store as big int in Cassandra
+			// gocql would cast big int to int64
+			return uint64(max.(int64)), nil
 		}
 	}
 	return 0, nil
